refactor: move weighted_utxo to seperate module
