+++ README.md
@@ -98,7 +98,7 @@ A basic performance comparison between implementations using commodity hardware 
 |implementation|pool size|ns/iter|
 |-------------:|---------|-------|
 |      Rust SRD|    1,000| 22,446|
-|      Rust BnB|    1,000|517,780|
+|      Rust BnB|    1,000|552,120|
 |  C++ Core BnB|    1,000|816,374|
 
 Note: The measurements where recorded using rustc 1.90 stable.  Expect worse performance with MSRV.
